+++ src/main/scala/com/lucendar/strm/common/inspect/CreateVidInspectTaskReq.java
@@ -1,5 +1,7 @@
 package com.lucendar.strm.common.inspect;
 
+import com.lucendar.strm.common.StreamingApi;
+
 import java.util.Arrays;
 import java.util.StringJoiner;
 
@@ -8,8 +10,25 @@ import java.util.StringJoiner;
  */
 public class CreateVidInspectTaskReq {
 
+    private String appId = StreamingApi.DEFAULT_APP_ID;
     private String simNo;
     private int[] channels;
+
+    /**
+     * 取 AppId
+     * @return AppId
+     */
+    public String getAppId() {
+        return appId;
+    }
+
+    /**
+     * 设置 AppId，可选，默认空白
+     * @param appId AppId
+     */
+    public void setAppId(String appId) {
+        this.appId = appId;
+    }
 
     /**
      * 取终端识别号
@@ -28,16 +47,16 @@ public class CreateVidInspectTaskReq {
     }
 
     /**
-     * 取通道ID列表
-     * @return 通道ID列表
+     * 取所要执行抓拍的通道ID列表
+     * @return 所要执行抓拍的通道ID列表
      */
     public int[] getChannels() {
         return channels;
     }
 
     /**
-     * 设置通道ID列表
-     * @param channels 通道ID列表
+     * 设置所要执行抓拍的通道ID列表
+     * @param channels 所要执行抓拍的通道ID列表
      */
     public void setChannels(int[] channels) {
         this.channels = channels;
@@ -46,6 +65,7 @@ public class CreateVidInspectTaskReq {
     @Override
     public String toString() {
         return new StringJoiner(", ", CreateVidInspectTaskReq.class.getSimpleName() + "[", "]")
+                .add("appId='" + appId + "'")
                 .add("simNo='" + simNo + "'")
                 .add("channels=" + Arrays.toString(channels))
                 .toString();