+++ src/main/scala/com/lucendar/strm/common/inspect/VidInspectTask.java
@@ -2,14 +2,33 @@ package com.lucendar.strm.common.inspect;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Objects;
 import java.util.StringJoiner;
 import java.util.stream.Collectors;
 
 /**
  * 视频抓拍巡检任务
  */
-public class VidInspectTask {
+public class VidInspectTask implements Cloneable {
+
+    /**
+     * 任务状态：0: 已登记
+     */
+    public static final int STATE__REGISTERED = 0;
+
+    /**
+     * 任务状态：1: 已创建，执行中
+     */
+    public static final int STATE__CREATED = 1;
+
+    /**
+     * 任务状态：2: 已结束
+     */
+    public static final int STATE__ENDED = 2;
+
+    /**
+     * 任务状态：-1: 任务创建失败
+     */
+    public static final int STATE__CREATE_FAILED = -1;
 
     private String taskId;
     private String instId;
@@ -18,6 +37,7 @@ public class VidInspectTask {
     private int[] channels;
     private String reqTm;
     private String endTm;
+    private int state;
     private int successCount;
 
     private List<VidInspectChannelState> states;
@@ -133,7 +153,7 @@ public class VidInspectTask {
 
     /**
      * 取任务创建时间
-     * @return 任务创建时间，，格式：yyyy-MM-dd HH:MM:SS
+     * @return 任务创建时间，格式：yyyy-MM-dd HH:MM:SS
      */
     public String getReqTm() {
         return reqTm;
@@ -164,35 +184,99 @@ public class VidInspectTask {
     }
 
     /**
-     * 取当前已成功的子任务数
-     * @return 当前已成功的子任务数
+     * 取任务状态
+     * <ul>
+     *     <li>0: 已登记</li>
+     *     <li>1: 已创建，执行中</li>
+     *     <li>2: 已结束</li>
+     *     <li>-1: 创建失败</li>
+     * </ul>
+     * @return 任务状态
+     */
+    public int getState() {
+        return state;
+    }
+
+    /**
+     * 设置任务状态
+     * <ul>
+     *     <li>0: 已登记</li>
+     *     <li>1: 已创建，执行中</li>
+     *     <li>2: 已结束</li>
+     *     <li>-1: 创建失败</li>
+     * </ul>
+     * @param state 任务状态
+     **/
+    public void setState(int state) {
+        this.state = state;
+    }
+
+    /**
+     * 取任务是否已处于结束态（已结束/创建失败）
+     * @return 任务是否已处于结束态（已结束/创建失败）
+     */
+    public boolean isFinalState() {
+        return state == STATE__ENDED || state == STATE__CREATE_FAILED;
+    }
+
+    /**
+     * 取当前已成功截图的通道数
+     * @return 当前已成功截图的通道数
      */
     public int getSuccessCount() {
         return successCount;
     }
 
     /**
-     * 设置当前已成功的子任务数
-     * @param successCount 当前已成功的子任务数
+     * 设置当前已成功截图的通道数
+     * @param successCount 当前已成功截图的通道数
      */
     public void setSuccessCount(int successCount) {
         this.successCount = successCount;
     }
 
     /**
-     * 取任务列表
-     * @return 任务列表
+     * 取通道执行状态列表
+     * @return 通道执行状态列表
      */
     public List<VidInspectChannelState> getStates() {
         return states;
     }
 
     /**
-     * 设置任务列表
-     * @param states 任务列表
+     * 设置通道执行状态列表
+     * @param states 通道执行状态列表
      */
     public void setStates(List<VidInspectChannelState> states) {
         this.states = states;
+    }
+
+    /**
+     * 是否所有通道都已经成功截图
+     *
+     * @return 是否所有通道都已经成功截图
+     */
+    public boolean isAllSuccess() {
+        return successCount == channels.length;
+    }
+
+    /**
+     * 取通道执行状态
+     *
+     * @param channel 通道号
+     * @return 通道执行状态
+     */
+    public VidInspectChannelState getChannelState(int channel) {
+        if (states == null)
+            return null;
+        else {
+            for (VidInspectChannelState state : states) {
+                if (state.getChan() == channel)
+                    return state;
+            }
+
+            return null;
+        }
     }
 
     @Override
@@ -203,10 +287,24 @@ public class VidInspectTask {
                 .add("appId='" + appId + "'")
                 .add("simNo='" + simNo + "'")
                 .add("channels=" + Arrays.toString(channels))
-                .add("reqTm=" + reqTm)
-                .add("endTm=" + endTm)
+                .add("reqTm='" + reqTm + "'")
+                .add("endTm='" + endTm + "'")
+                .add("state=" + state)
                 .add("successCount=" + successCount)
-                .add("states=" + (states != null ? states.stream().map(Objects::toString).collect(Collectors.joining(", ")) : "null"))
+                .add("states=" + states)
                 .toString();
+    }
+
+    @Override
+    public VidInspectTask clone() {
+        try {
+            VidInspectTask r = (VidInspectTask) super.clone();
+            if (states != null) {
+                r.setStates(states.stream().map(VidInspectChannelState::clone).collect(Collectors.toList()));
+            }
+            return r;
+        } catch (CloneNotSupportedException e) {
+            throw new RuntimeException(e);
+        }
     }
 }